edited file structure
