+++ Dockerfile
@@ -1,12 +1,12 @@
 # syntax=docker/dockerfile:1
 
-FROM python:3.8
+FROM python:3.8-slim-buster
 
-WORKDIR /zest
+WORKDIR /
 
 COPY requirements.txt requirements.txt
 RUN pip3 install -r requirements.txt
 
 COPY . .
 
-CMD [ "python", "main.py"]
+CMD [ "python3", "main.py"]
