--- login.py
@@ -1,32 +1,0 @@
-from dotenv import load_dotenv
-import sys
-import os
-import dxpy as dx
-
-load_dotenv()
-
-def dx_login():
-
-    if len(sys.argv) > 1:
-        # config / env file passed as arg
-        # read in to dict and don't touch env variables
-        AUTH_TOKEN = os.getenv("AUTH_TOKEN")
-    else:
-        # try to get auth token from env (i.e. run in docker)
-        try:
-            AUTH_TOKEN = os.environ["AUTH_TOKEN"]
-        except NameError as e:
-            raise NameError(
-                'auth token could not be retrieved from environment and no .env file passed'
-            )
-
-    print(f'Authentication Token: {AUTH_TOKEN}')
-
-    # env variable for dx authentication
-    DX_SECURITY_CONTEXT = {
-        "auth_token_type": "Bearer",
-        "auth_token": AUTH_TOKEN
-    }
-
-    # set token to env
-    dx.set_security_context(DX_SECURITY_CONTEXT)
